fix: resolve type mismatches in actions.c and players.c
- actions.c: change val_28, val_30, val_24, val_34 from int to short
  to match encounter_modify_pursuit_desires signature (short*)
- players.c: cast record[2] to (const char*) for FUN_00058c40 call

+++ src/halo/ai/actions.c
@@ -2740,10 +2740,10 @@ char actor_action_handle_lost_contact(int actor_handle)
   char found;
   char flag_2c;
   char flag_38;
-  int val_24;
-  int val_28;
-  int val_30;
-  int val_34;
+  short val_24;
+  short val_28;
+  short val_30;
+  short val_34;
   int threat;
   short firing_pos;
   short threshold;
